requirements.txt
beepy
pandas
googletrans
python-dotenv